refactor: 랜덤생성기 패키지 util -> model 변경
