+++ src/main/java/model/NumberGenerator.java
@@ -1,4 +1,4 @@
-package util;
+package model;
 
 public interface NumberGenerator {
     int generateNumber();
