+++ src/main/java/model/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package util;
+package model;
 
 public class RandomNumberGenerator implements NumberGenerator {
     private static final int BOUND = 10;
