+++ src/test/java/controller/RaceTest.java
@@ -6,7 +6,7 @@ import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.PrintStream;
 import org.junit.jupiter.api.Test;
-import util.NumberGenerator;
+import model.NumberGenerator;
 
 class RaceTest {
